\name{DefaultColorSequence}
\alias{DefaultColorSequence}
\docType{data}
\title{
Default color sequence for plots
}
\description{
Defines the default color sequence for plots made within the Projections package.
}
\usage{data("DefaultColorSequence")}
\format{
  A vector with 562 different strings describing colors for plots.
}