\name{Hepta}
\alias{Hepta}

\docType{data}
\title{
Hepta form FCPS
}
\description{
clearly defined clusters, different variances
}
\usage{data("Hepta")}

\details{
Size 212, Dimensions 3, stored in \code{Hepta$Data}

Classes 7, stored in \code{Hepta$Cls}
}

\references{
[Ultsch,2005]   Ultsch, A.:Clustering with SOM: U*C, In Proc. Workshop on Self-Organizing Maps, Paris, France, pp. 75-82, 2005.
}
\examples{
data(Hepta)
str(Hepta)
}
\keyword{Hepta}
\keyword{dataset}
\keyword{FCPS}