\name{Hepta}
\alias{Hepta}

\docType{data}
\title{
Hepta is part of the Fundamental Clustering Problem Suit (FCPS) [Thrun/Ultsch, 2020].
}
\description{
clearly defined clusters, different variances
}
\usage{data("Hepta")}

\details{
Size 212, Dimensions 3, stored in \code{Hepta$Data}

Classes 7, stored in \code{Hepta$Cls}
}

\references{
[Thrun/Ultsch, 2020]	Thrun, M. C., & Ultsch, A.: Clustering Benchmark Datasets Exploiting the Fundamental Clustering Problems, Data in Brief,Vol. 30(C), pp. 105501, DOI 10.1016/j.dib.2020.105501 , 2020. 
}
\examples{
data(Hepta)
str(Hepta)
}
\keyword{Hepta}
\keyword{dataset}
\keyword{FCPS}